Update README.md
Breadcrumb for Magento 2, when products are in multiple categories. using ui Grid to solve this problem.

Admin path: stores > configuration > Magentoyan (Tab) > Breadcrumb .
You can enable/disable the module.
There is a link to a Grid "priority of the categories", click this link and edit the grid rows, "Position" and "Is Active" columns are
editabele. position should be int numbers, priority is given to the smaller number.